Fixed multiple log handlers
During hyperparameter optimisation of a hiclass classifier, HierarchicalClassifier._create_logger() adds a new StreamHandler to the logger every time it is instantiated. This means that on the nth iteration, every log line in the console is repeated n times.

By checking if a hander is already set on the logger, we can skip adding another one when we call _create_logger()